requirements.txt
astroid==2.3.2
beautifulsoup4==4.8.1
bootstrap4==0.1.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.6
django-appconf==1.0.3
django-autocomplete-light==3.4.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.0.1
django-countries==5.5
django-crispy-forms==1.8.0
django-cruds-adminlte==0.0.16
django-image-cropping==1.2.0
django-modeltranslation==0.13.4
django-rosetta==0.9.3
django-settings-export==1.2.1
django-taggit==1.1.0
django-userforeignkey==0.3.0
djangoajax==3.1
djangorestframework==3.10.3
easy-thumbnails==2.6
html5lib==1.0.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pep8==1.7.1
Pillow==6.2.1
pipenv==2018.11.26
polib==1.1.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.3
pylint-django==2.0.11
pylint-plugin-utils==0.6
PyPDF2==1.26.0
pytz==2019.3
reportlab==3.5.31
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.7.7
virtualenv-clone==0.5.3
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
